camera/auth/requirements.txt
behave==1.2.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dbus-python==1.3.2
idna==3.4
parse==1.19.0
parse-type==0.6.0
pycairo==1.21.0
pycparser==2.21
PyGObject==3.42.2
python-networkmanager==2.2
requests==2.28.1
responses==0.22.0
six==1.16.0
toml==0.10.2
types-toml==0.10.8
urllib3==1.26.12